requirements/production.txt
# Install everything that needs to be compiled
-r compiled.txt

# Django stuff
Django>=1.7

# Templates
django_compressor

# Security
bleach
#cef

# Logging
#raven

# Celery: Message queue
celery
django-celery

# L10n/i18n
#Babel
#-e git://github.com/clouserw/tower.git#egg=tower

# Caching
python-memcached

# Admin
django-debug-toolbar

# Migrations

# Feature testing
#django-waffle

# Imaging Library
Pillow

# Segmentation and Classification
numpy
matplotlib
mahotas
pandas

